copr: Add get_rand util fn
